requirements.txt
amqp==2.4.2
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.11
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-redis==4.10.0
django-reversion==3.0.3
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.4.2
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
itsdangerous==1.1.0
jdcal==1.4
kombu==4.5.0
odfpy==1.4.0
openpyxl==2.6.2
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2018.9
PyYAML==5.1
redis==3.2.1
requests==2.21.0
six==1.12.0
tablib==0.13.0
urllib3==1.24.1
vine==1.3.0
xadmin==0.6.1
xlrd==1.2.0
xlwt==1.3.0
